==> 00-preface.Rmd <==
---
title: "AMMonitor: Remote Monitoring of Biodiversity in an Adaptive Framework"
output:
  html_document:
  md_document:
    variant: markdown_github
    toc: true
    toc_depth: 3
    includes:
      in_header: header.html
bibliography: refs.bib
csl: plos.csl
---

```{r, echo = F}
# to create md file for the code.usgs.gov wiki, in the console type:
# rmarkdown::render(input = "00-preface.RMD", output_format = "md_document", output_file = "Chapter-00-Motivation-and-Gratitude.md")

```




```{r echo = F}
source("includes.R")
```


# Chapter Introduction

Amid climate change and rapidly shifting land uses, effective methods for monitoring natural resources are critical to support scientifically-informed resource management decisions [@Holling1978; @Walters1986; @Lee1993; @Pollock2002; @Allen2015]. The practice of using Autonomous Monitoring Units (AMUs) to monitor wildlife species has grown immensely in the past decade, with monitoring projects across species from birds, to bats, amphibians, insects, terrestrial mammals, and marine mammals [@August2015; @Burton2015].

AMUs have many benefits. Primarily, they can be deployed for long periods of time to collect massive amounts of data, such as audio recordings and photos. Having a record of audio and photo data allows researchers to carefully verify and analyze species identifications *a posteriori* [@Hobson2002]. 

However, automated methods have several limitations. First, individual AMUs can be expensive, running over $800 USD for commercial devices [@WildlifeAcoustics2019], although cost-effective models are becoming more common [@Whytock2017; @Hill2018]. Second, data is typically stored on AMUs until researchers can retrieve it, causing time lapses between data collection, analysis, and results. Such delays hamper the ability to efficiently address pressing ecological challenges and track progress toward management objectives. Third, the data management requirements of an AMU research effort are often immense. A monitoring program is a collection of people, equipment, monitoring locations, location characteristics, research objectives, and data files, with multiple moving parts to manage. Without a comprehensive framework for efficiently moving from raw data collection to results and analysis, monitoring programs are limited in their capacity to characterize ecological processes and inform management decisions [@Gregory2006; @Rehme2011; @Fontaine2011; @Greig2013; @Rist2013; @Fischman2016; @Williams2016]. 

**AMMonitor** is an open source R package dedicated to collecting, storing, and analyzing AMU information in a way that 1) is cost-effective, 2) can efficiently process and store information, and 3) can take advantage of the vast and growing community of R analytics. We created **AMMonitor** for the Bureau of Land Management to monitor high priority wildlife across the southern California Solar Energy Zone (SEZ), including the Couch’s Spadefoot (*Scaphiopus couchii*), kit fox (*Vuples macrotis*), coyote (*Canis latrans*), and a variety of bird species, such as the Verdin (*Auriparus flaviceps*), Black-tailed Gnatcatcher (*Polioptila melanura*), and Eurasian Collared-Dove (*Streptopelia decaocto*). The agency has established management objectives (benchmarks) to ensure the persistence of sensitive species and minimize the spread of invasive species across the SEZ as solar energy projects are added to the landscape. In developing **AMMonitor**, our primary goal was to create a system for handling and processing massive amounts of data to allow BLM to quickly ascertain species distribution patterns (e.g., an occupancy analysis) in relation to their management objectives. 

In broad terms, the **AMMonitor** approach starts with ecological hypotheses or natural resource management objectives (Figure 0.1; boxed). Data are collected with Autonomous Monitoring Units (AMUs) to test hypotheses or to evaluate the state of a resource with respect to a management objective. Acoustic recordings and photos are collected and delivered to the cloud. Raw and processed data are stored in a SQLite database. The data can be analyzed with a wide variety of analytical methods, often models of abundance or occupancy pattern. These analyses can be stored, and resulting outputs can be compared with research and monitoring objectives to track progress toward management goals. The final results are assessed with respect to hypotheses or objectives. Thus, the **AMMonitor** package places the monitoring data into an **a**daptive **m**anagement framework [@Williams2011].

<kbd>

```{r, eval = T, echo = F, out.width = '100%', echo = F, fig.align = 'center'}

knitr::include_graphics('Chap0_Figs/fig1.png', dpi = 500)
```
</kbd>

>*Figure 0.1. The general AMMonitor framework begins with basic research hypotheses or applied resource management objectives (boxed).*

The **AMMonitor** approach was developed with a prototype of 20 smartphone-based AMUs [@BalanticStatistical; @BalanticOccupancy; @BalanticTemporal]. Since then, we have added the capacity to use the smartphone’s camera by enabling timed photographs as well as motion-triggered photographs, allowing the smartphones to act as camera traps. However, the **AMMonitor** approach does not require the use of smartphones. Its flexibility allows the analyis of data collected by other autonomous devices, and further permits the storage of results from other analytical systems for additional processing in R. 

This guide provides step-by-step instructions for using **AMMonitor** in its current form for monitoring programs that rely on remotely-captured data for use in adaptive management. We welcome collaborators who may be interested in improving or building on our approach.


# Gratitude

We thank Mark Massar and the Bureau of Land Management for essential field support and guidance, Jon Katz and Jim Hines for programming assistance, and John Sauer for critical review of the software and documentation. 



# Chapter References
